backend/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-cors==5.0.1
flask-restful==0.3.10
flask-swagger-ui==4.11.1
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.0.0
marshmallow==3.20.2 